add audit script to Makefile
The auto-import handling by vsCode is very useful, most of the time.
However, it sometimes uses the wrong version of go-gl; this can cause
numerous problems -- typically the code will compile, but then crash
with a segfault or similar. This usually leads to wasted time trying
to identify what has gone wrong.

This script should catch that for us.